+++ Source/NSIndexPath+Etoile.m
@@ -77,7 +77,7 @@ Will raise an NSInvalidArgumentException if the separator is nil. */
         return @"";
 
     NSString *path = [NSString stringWithFormat: @"%lu", (unsigned long)[self firstIndex]];
-    int indexCount = [self length];
+    NSUInteger indexCount = [self length];
 
     for (int i = 1; i < indexCount; i++)
     {
